requirements.txt
python-dotenv~=0.20.0
Pyrogram~=2.0.30
tgcrypto~=1.2.3
tortoise-orm~=0.19.0
aerich~=0.7.0